Refactor : 폴더 업데이트

+++ src/main/java/com/encore/bbs/board/dto/CountryDto.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter @Setter
 @NoArgsConstructor
 public class CountryDto {
-    private Long countryId;
+    private int countryId;
     private String continentName;
     private String countryName;
